RDM-3916 refactored class management
