+++ src/shared/components/create-case-filters/create-case-filters-selection.model.ts
@@ -1,0 +1,11 @@
+export class CreateCaseFiltersSelection {
+    jurisdictionId: string
+    caseTypeId: string
+    eventId: string
+
+    constructor() {
+        this.jurisdictionId = 'TEST',
+            this.caseTypeId = 'TestAddressBookCase',
+            this.eventId = 'startCase'
+    }
+}
