+++ src/shared/components/create-case-filters/index.ts
@@ -1,1 +1,2 @@
 export * from './create-case-filters.module';
+export * from './create-case-filters-selection.model';
